Refactor part list
